Recreate does not work as expected either
I changed the reconnect test to be extremely obvious as to why it does not work, you should get an error if you never successfully connected before and then you reconnect, but you don't because of the condition inside reconnect to only reconnect if connected. As I said before, I think that reconnect should try to reconnect regardless of whether or not you are currently connected. In fact, if I were to put a condition in there, I would make it return if it were already connected, not if it wasn't. Also, the reconnect and recreate functions should not modify the status property because that is set in the keepalive function.

The Recreate function has the exact same problem so the test fails when you call recreate because the old session is the exact same object as the new session.